Fast forward update for Hello Kendo UI Part 2. Shame on me.

+++ hello-kendo-ui-part-2/hello-kendo-ui/Controllers/EmployeesController.cs
@@ -11,7 +11,6 @@ namespace hello_kendo_ui.Controllers {
     // Inheriting from the APIController class will designate this as a WebAPI endpoint
     public class EmployeesController : ApiController {
 
-
         // the linq to sql context that provides the data access layer
         private Data.NorthwindContextDataContext _context = new Data.NorthwindContextDataContext();
         // the current http request being made
@@ -19,8 +18,6 @@ namespace hello_kendo_ui.Controllers {
 
         // WebAPI will respond to an HTTP GET with this method
         public Models.Response Get() {
-
-            System.Threading.Thread.Sleep(1000);
 
             // the the take and skip parameters off of the incoming request
             int take = _request["take"] == null ? 10 : int.Parse(_request["take"]);
@@ -31,16 +28,11 @@ namespace hello_kendo_ui.Controllers {
             // defined model objects for easy serialization. skip and then
             // take the appropriate number of records for paging.
             var employees = (from e in _context.Employees
-                             select new Models.Employee {
-                                 Id = e.EmployeeID,
-                                 FirstName = e.FirstName,
-                                 LastName = e.LastName
-                             }).Skip(skip).Take(take).ToArray();
+                             select new Models.Employee(e)).Skip(skip).Take(take).ToArray();
 
             // returns the generic response object which will contain the 
             // employees array and the total count
             return new Models.Response(employees, _context.Employees.Count());
-
         }
 
         public HttpResponseMessage Post(int id) {
@@ -49,41 +41,70 @@ namespace hello_kendo_ui.Controllers {
             var response = new HttpResponseMessage();
 
             try {
+                // select the employee from the database where the id
+                // matches the one passed in at api/employees/id
                 var employeeToUpdate = (from e in _context.Employees
                                         where e.EmployeeID == id
                                         select e).FirstOrDefault();
 
-                if (employeeToUpdate != null && _request["LastName"] != null) {
-                    employeeToUpdate.LastName = _request["LastName"];
+                // if there was an employee returned from the database
+                if (employeeToUpdate != null) {
+                    
+                    // update the employee object handling null values or empty strings
+                    employeeToUpdate.LastName = string.IsNullOrEmpty(_request["LastName"]) ? employeeToUpdate.LastName : _request["LastName"];
+                    employeeToUpdate.Address = string.IsNullOrEmpty(_request["Address"]) ? employeeToUpdate.Address : _request["Address"];
+                    employeeToUpdate.City = string.IsNullOrEmpty(_request["City"]) ? employeeToUpdate.City : _request["City"];
+                    employeeToUpdate.BirthDate = string.IsNullOrEmpty(_request["BirthDate"]) ? employeeToUpdate.BirthDate : Convert.ToDateTime(_request["BirthDate"]);
+
+                    // submit the changes to the database
                     _context.SubmitChanges();
 
+                    // set the server response to OK
                     response.StatusCode = HttpStatusCode.OK;
 
                 } else {
+                    // we couldn't find the employee with the passed in id
+                    // set the response status to error and return a message
+                    // with some more info.
                     response.StatusCode = HttpStatusCode.InternalServerError;
                     response.Content = new StringContent(string.Format("The employee with id {0} was not found in the database", id.ToString()));
                 }
             } catch (Exception ex) {
+                // something went wrong - possibly a database error. return a
+                // 500 server error and send the details of the exception.
                 response.StatusCode = HttpStatusCode.InternalServerError;
                 response.Content = new StringContent(string.Format("The database updated failed: {0}", ex.Message));
             }
 
+            // return the HTTP Response.
             return response;
         }
 
-        // WebAPI will respond to an HTTP DELETE with this method
-        public void Delete(int id) {
+        public Models.Response Delete(int id) {
+            try {
+                // retrieve the employee to update from the database
+                // based on the parameter passed in from api/employees/id
+                var employeeToDelete = (from e in _context.Employees
+                                        where e.EmployeeID == id
+                                        select e).FirstOrDefault();
 
-            var employeeToDelete = (from e in _context.Employees
-                                    where e.EmployeeID == id
-                                    select e).SingleOrDefault();
+                // if a valid employee object was found by id
+                if (employeeToDelete != null) {
+                    // mark the object for deletion
+                    _context.Employees.DeleteOnSubmit(employeeToDelete);
+                    // delete the object from the database
+                    _context.SubmitChanges();
 
-            // mark the object for delete in the context
-            _context.Employees.DeleteOnSubmit(employeeToDelete);
-
-            // commit the change
-            _context.SubmitChanges();
-
-        }
+                    // return an empty Models.Response object (this returns a 200 OK)
+                    return new Models.Response();
+                } else {
+                    // otherwise set the error field of a response object and return it.
+                    return new Models.Response(string.Format("The employee with id {0} was not found in the database", id.ToString()));
+                }
+            } catch (Exception ex) {
+                // something went wrong. set the errors field of 
+                return new Models.Response(string.Format("There was an error updating employee with id {0}: {1}", id.ToString(), ex.Message));
+            }
+        }           
     }
 }
